removes table_type.py since it is not used in the e2e tests.
