+++ cloudbuild/python-scripts/utils/utils.py
@@ -5,28 +5,77 @@ import datetime
 import os
 import re
 import avro
+import random
+import string
 import avro.datafile
 import avro.io
 from google.cloud import bigquery
+
+
+def is_perfect_hour(datetime_obj):
+    """Returns True if the datetime object is a perfect hour, False otherwise."""
+    return (
+        datetime_obj.minute == 0
+        and datetime_obj.second == 0
+        and datetime_obj.microsecond == 0
+    )
+
+
+def generate_string():
+    return ''.join(
+        random.choices(string.ascii_letters, k=random.randint(8, 10))
+    )
+
+
+def generate_long():
+    return random.choice(range(0, 10000000))
+
+
+def generate_timestamp(current_timestamp):
+    """Method to generate a random datetime within the given hour.
+    Args:
+      current_timestamp: Date is generated within one hour of this timestamp.
+    Returns:
+      datetime object: Containing generated timestamp.
+    """
+    next_hour = current_timestamp + datetime.timedelta(hours=1)
+    random_timestamp = random.randint(
+        int(current_timestamp.timestamp()), int(next_hour.timestamp())
+    )
+    utc = datetime.timezone.utc
+    random_timestamp_utc = datetime.datetime.fromtimestamp(
+        random_timestamp, utc
+    )
+    # Check if the generated entry is a perfect hour.
+    # Note: It is only for the case of hour based partitioning
+    # (as created in our test).
+    # If the values inserted are changed to DAY or any other partitioning,
+    # the prevention of borderline entry generation needs to be
+    # changed accordingly.
+    while is_perfect_hour(random_timestamp_utc):
+        # Keep on regenerating.
+        random_timestamp = random.randint(
+            int(current_timestamp.timestamp()), int(next_hour.timestamp())
+        )
+        utc = datetime.timezone.utc
+        random_timestamp_utc = datetime.datetime.fromtimestamp(
+            random_timestamp, utc
+        )
+    return random_timestamp_utc
 
 
 class TableCreationUtils:
     """Class containing all the basic utilities for creation of BQ table.
 
     Attributes:
-      table_type: Table type to be created: can be 'large_table',
-        'large_row_table', 'complex_schema_table', 'unbounded_table'
       schema: The table schema to be inserted.
       number_of_rows_per_batch: Number of rows per thread.
-      avro_file_local: The name of the local avro file to generate the data to.
       table_id: ID of the table of the form {project_id}.{dataset_id}.{table_id}
     """
 
     def __init__(
         self,
         simple_avro_schema_string,
-        table_type,
-        avro_file_local,
         number_of_rows_per_batch,
         table_id,
     ):
@@ -34,16 +83,31 @@ class TableCreationUtils:
 
         Args:
           simple_avro_schema_string: Schema of the table in avro format.
-          table_type: `TableType` Object. For custom schema and record generation.
-          avro_file_local: The name of the local avro file to generate the data to.
           number_of_rows_per_batch: Number of rows per thread.
           table_id: ID of the table of the form {project_id}.{dataset_id}.{table_id}
         """
-        self.table_type = table_type
         self.schema = avro.schema.parse(simple_avro_schema_string)
         self.number_of_rows_per_batch = number_of_rows_per_batch
-        self.avro_file_local = avro_file_local
         self.table_id = table_id
+
+    def write_rows(
+        self,
+        number_of_rows_per_batch,
+        writer,
+        partition_number,
+        current_timestamp,
+    ):
+        offset_timestamp = current_timestamp + datetime.timedelta(
+            hours=partition_number
+        )
+
+        # Write the specified number of rows.
+        for _ in range(number_of_rows_per_batch):
+            writer.append({
+                'name': generate_string(),
+                'number': generate_long(),
+                'ts': generate_timestamp(offset_timestamp),
+            })
 
     def write_avros(self, avro_file_local_identifier, partition_number, current_timestamp):
         """Method to generate fake records for BQ table.
@@ -66,7 +130,7 @@ class TableCreationUtils:
             avro.io.DatumWriter(),
             self.schema,
         )
-        self.table_type.write_rows(
+        self.write_rows(
             self.number_of_rows_per_batch,
             writer,
             partition_number,
